auth/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.3
distlib==0.3.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
fastapi-cli==0.0.5
filelock==3.12.4
gitdb==4.0.11
GitPython==3.1.31
graphviz==0.20.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
Jinja2==3.1.4
lark==1.1.8
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.3
pandas==2.1.4
phonenumbers==8.13.43
platformdirs==3.11.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.10
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-hcl2==4.3.0
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
rich==13.7.1
setuptools==65.6.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.2
tqdm==4.65.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.24.5
watchfiles==0.23.0
websockets==13.0
